Fix autoload for non-underscored libraries
When a constant class name or module name is included as part of the "explicitly_unloadable_constants", it is assumed that the files that relate to that class are named according to the rules of camel to underscore convention, e.g. "OmniAuth" -> "lib/omni_auth.rb".  However, some libraries, like "OmniAuth", do not follow this convention.  OmniAuth uses "lib/omniauth.rb", and this causes the reload process to fail catastrophically.  Dependencies will happily unload the module/class, but will fail to reload it as Dependencies can not match the file name.  This patch simply assumes that the library name may not contain underscores.